Add "hardcodedGadgets64bit.c" with assembly instruction sequences for assignment gadgets and arithmetical operations gadgets. I will add more gadgets in future commits.
It would perhaps be better to find legitimate instruction sequences in libc, the vulnerable binary or elsewhere through manual (or maybe automatic) analysis,
but for now we will go with this helper file since we are interested in the ROP-code-to-payload automatisation step, not the gadget-finding automatisation step.
Also, this file will be convenient to work with.