Make the static methods of "InstructionConverter" into instance methods and initialize the Keystone and Capstone objects just once in the instance constructor.
This was done for the sake of optimization and the hope of a decent improvement in execution time when building the inner trie of "VirtualMemoryInfo".

However, the time difference doesn't seem to be noticeable :/.
The explanation for that seems to be that the Capstone functions (e.g. cs_open(), cs_option(), cs_close()) are *very* fast.

Warning: The Keystone functions (e.g. ks_open(), ks_option(), ks_close()) are *not* fast. But this didn't affect the performance since we call Cornerstone a lot, but we don't call Keystone much.

These are some time measurements done locally:
	Measuring Keystone:
	TEST_COUNT = 10000
		"PUSH rbp; MOV rax, qword ptr [rip + 0x13b8]" -> 0x55 0x48 0x8B 0x05 0xB8 0x13 0x00 0x00
		Compiled: 8 bytes, statements: 2
	ks_open() took 0.0005217269 seconds!
	ks_option() took 0.0000476804 seconds!
	ks_asm() took 0.0005436751 seconds!
	ks_free() took 0.0000487973 seconds!
	ks_close() took 0.0000625083 seconds!

	Measuring Capstone:
	TEST_COUNT = 10000
		0x1000:	push		rbp
		0x1001:	mov		rax, qword ptr [rip + 0x13b8]
	cs_open() took 0.0000000037 seconds!
	cs_option() took 0.0000000037 seconds!
	cs_disasm() took 0.0000000044 seconds!
	cs_free() took 0.0000000036 seconds!
	cs_close() took 0.0000000036 seconds!