Optimize the construction of the inner trie of the "VirtualMemoryInfo" class by using the prefix-free nature of the x86 architecture.
The time went from ~41 seconds to ~4.5 seconds.

The idea is that, given an index "first", the recursive algorithm might try, during the construction of the trie,
to disassemble multiple segments starting at first (e.g. [first, last1], [first, last2], ...).
But, given that x86 is a prefix-free architecture, there is only one possible [first, last] segment that can be valid.
Thus, when we are asked to analyze [first, something], we check for the actual valid segment (if it exists) and store this information.
In this way, we call Capstone only once for each possible "first" index, even if we analyze multiple [first, something] segments.