Add code to print the instruction prefix bytes as well in the `InstructionConverter::printCapstoneInformationForInstructions()` method.
Also, remove the statement which asserted that the number of instructions parsed from the input string by Keystone is the same as the number of instructions decoded back from the bytes by Capstone. It seems that Keystone parses prefix bytes (or at least the "REP"/0xF3 prefix byte) as a separate instruction, which makes the assert wrongfully fail.